python/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deptrast"
version = "3.0.1"
description = "The ultimate dependency tree converter, enhancer, and streamliner"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Jeff Williams", email = "jeff.williams@contrastsecurity.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "requests>=2.28.0",
    "cyclonedx-python-lib>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
]

[project.scripts]
deptrast = "deptrast.__main__:main"

[project.urls]
Homepage = "https://github.com/planetlevel/deptrast"
Repository = "https://github.com/planetlevel/deptrast"
